All pathways fully done. Polyline in med map deleted. Code cleaned.
