+++ app/src/main/res/layout/activity_short_map.xml
@@ -6,5 +6,5 @@
     android:name="com.google.android.gms.maps.SupportMapFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.wbunkry.activities.ShortMapActivity" />
+    tools:context="com.example.wbunkry.activities.ShortPathActivity" />
